Add support for links in tiptap
