+++ apps/wiki/package.json
@@ -22,6 +22,7 @@
     "@iconify/react": "^4.1.1",
     "@tiptap/extension-color": "^2.2.4",
     "@tiptap/extension-image": "^2.2.4",
+    "@tiptap/extension-link": "^2.2.4",
     "@tiptap/extension-list-item": "^2.2.4",
     "@tiptap/extension-text-style": "^2.2.4",
     "@tiptap/pm": "^2.2.4",
