+++ apps/wiki/src/app/page.tsx
@@ -8,8 +8,6 @@ export default function Page() {
           access={true}
           json='
           <p>This is a basic example of implementing images. Drag to re-order.</p>
-          <img src="https://source.unsplash.com/8xznAGy4HcY/800x400" />
-          <img src="https://source.unsplash.com/K9QHL52rE2k/800x400" />
           '
         />
       </div>
